Add delete flag to clean up endpoint

+++ activiti-cloud-services-query/activiti-cloud-services-query-rest/src/main/java/org/activiti/cloud/services/query/rest/ProcessInstanceCleanUpController.java
@@ -42,13 +42,15 @@ public class ProcessInstanceCleanUpController {
 
     @RequestMapping(method = RequestMethod.DELETE, path = "/export")
     public PagedResources<ProcessInstanceResource> exportProcessInstances (@QuerydslPredicate(root = ProcessInstanceEntity.class) Predicate predicate,
-                                                                           Pageable pageable) {
+                                                                           Pageable pageable,
+                                                                           boolean delete) {
 
         PagedResources<ProcessInstanceResource> result = pagedResourcesAssembler.toResource(pageable,
                 processInstanceRepository.findAll(predicate,
                         pageable),
                 processInstanceResourceAssembler);
 
+        if(delete)
         processInstanceRepository.deleteAll();
 
         return result;
